EdgeSolution/modules/system-monitor/requirements.txt
azure-iot-device==2.12.0
pytz==2024.1